fixed a bug with incorrect calculation of the query and tuned query time and made the query rac aware
In the query, the max query length and the tuned undo retention are converted to a "days hours:minutes:seconds" format using the extract function.
This means that the value must be of the timestamp datatype and not integer.
To convert the original integer value to a timestamp, the following trick was used: systimestamp + integer/60/60/24 - systimestamp
However, because of implicit data type conversion when adding an integer to a timestamp value, the timezone information on the first systimestamp was lost (as the resulting datatype is date, not timestamp). This gives a wrong result when the second systimestamp value is subtracted from the result.

To fix this, the integer was converted to an interval first: systimestamp + numtodsinterval(integer, 'seconds') - systimestamp

The query was also made rac aware by switching to gv$undostat and adding the inst_id column to the relevant places

query formatting was changed for better readability