fixed a bug with incorrect calculation of the elapsed and remaining time in the longops2.sql script
In the query, the elapsed and remaining are converted to a "days hours:minutes:seconds" format using the extract function.
This means that the value must be of the timestamp datatype and not integer.
To convert the original integer value to a timestamp, the following trick was used: systimestamp + integer/60/60/24 - systimestamp
However, because of implicit data type conversion when adding an integer to a timestamp value, the timezone information on the
first systimestamp was lost (as the resulting datatype is date, not timestamp).
This gives a wrong result when the second systimestamp value is subtracted from the result.

To fix this, the integer was converted to an interval first: systimestamp + numtodsinterval(integer, 'seconds') - systimestamp