fixed ordering bug in rmancat_all_available_backupsets.sql
Because the formatted representation of the subjob_start_time column had the same alias as the original column,
the ordering on the job start time was done on the formatted representation instead of the real date value.
Causing a list like 01/11/2015, 01/12/1025, 02/11/2015.

This was fixed by prefixing the sort column with the table name, after which the date value is used for the sort.

To avoid further confusion all columns are now prefixed with the table name / alias