Add show_cbo_stats12c.sql script
Add the show_cbo_stats12c.sql script.
This script gives an extensive overview of the cbo statistics in place for a given table (table, partition / subpartition, column),
the statistic extensions, on the indexes for this table and lists the cbo directives related to this table.

The user executing this statement needs select privileges on the following views:

  - dba_tab_statistics
  - dba_stat_extensions
  - dba_tab_columns
  - dba_tab_col_statistics
  - dba_tab_partitions
  - dba_part_col_statistics
  - dba_tab_subpartitions
  - dba_subpart_col_statistics
  - dba_indexes
  - dba_ind_columns
  - dba_ind_expressions
  - dba_part_indexes
  - dba_ind_statistics
  - dba_sql_plan_dir_objects
  - dba_sql_plan_directives

With the exception of the last 2 views related to the plan directives, these views can be replaced by their all_ counterparts

When running the script, it will ask for the table owner (defaults to the current user) and the table_name.