Fix mixup between begin and end interval time
The column bsnap_time actually contained the end time of the period.
This fix changes the select to list the begin of the snapshot period and the end of the snapshot period, as wel as the actuall snapshot id's.
These id's make it easier to later on create the full awr report.