Add reset_expired_users.sql
The reset_expired_users.sql script reset the password of expired users to their current value.
This to open the account again without having the actually change the password
Accepts a username (with wildcards) as input parameter. If no input parameter was given, it will run for all users with status expired / expired(grace)
Note that if in the profile a password_reuse_max is set, the profile for that user needs to be changed before this procedure will work

Need to run as a sysdba user
Passwords of common users need to be changed on CDB level, PDB users from within the PDB

Tested on 12cR1